requirements.txt
torch
transformers
datasets
sacrebleu
aim
sentencepiece
hydra-core